WARNING! WARNING! DANGER, WILL ROBINSON!
Make missing DOCTYPE a non-fatal-error, emitting instead a long explanation
of why this is bad and how to fix it. This is an experiment to attempt to
address the issues brough up on w-v by Howcome and Bj?rn. It may get
backed out before release if feedback indicates that to be the better
compromise!