First pass at integrating validator with an external checker.
Enabling the external checker is done in the validator's configuration.

If enabled and set up, the validator will, in specific cases, call
(via HTTP + API) the external checker instead of its own internal parsers.

The first external checker used is the validator.nu html5 parser and checker.
For the time being the checker is called only when the user specifies the
(X)HTML5 doctype in the options. No automatic swiching on <!DOCTYPE html> at
the moment - although that would be feasible -.

TBD: location (line, column) of errors (seems like there is a bug in the
current XML output of the html5 checker), warning/info distinction (missing
test cases to trigger either) and forwarding of the error message explanations.