Fix tidy output with "wide" characters (#7592).
Strangely, at least with Fedora 11's perl 5.10.0-73.fc11.x86_64, Encode 2.23,
Encode::decode_utf8(...) is documented to be equivalent to
Encode::decode("utf8", ...) but for some inputs the former has no problems
with, the latter barfs "Cannot decode string with wide characters" or results
in corrupt conversions.  So switch to using the former.