i18n enhancements:
Currently, template directories are found in the order listed in the
Language option in validator.conf.

Enhanced patches no longer use the Language option.

By referring to the Accept-Languages header value of the browser, the
template directory is found in Accept-Language order.

For examples:

For example, if Accept-Languages is:

```
Accept-Langauge: ko-KR,ko:q=1.0,en-GB:q=0.7,en:q=0.3
```

The validator will find the template in the following order:

1. template/ko_KR
1. template/ko
1. tempalte/en_GB
1. template/en

If there is no template corresponding to Accept-Languages, template/en_US will be used.